+++ src/DatabasesManager/MySQLManager.java
@@ -81,7 +81,7 @@ public class MySQLManager {
                 } 
                 
                 //remove compiling file if didn't exist before
-            	if (remove){
+            	if (remove&&!PropertiesHandler.CompileFile){
 	        		if(file1.delete()){
 	        			System.out.println("Compilation file " + file1.getName() + " has been deleted!");
 	        		}else{
